implement back jump

+++ src/animation.hpp
@@ -10,7 +10,7 @@
 
 class Animation{
 public:
-	Animation(Man &m, std::vector<Stone> &s,Bonus &b, int jump_ongoing=0, double jumped=0.0, int anim_stone=0, double stone=0.0, int num = -1,  int pom_anim=0,float score_num=0.0)
+	Animation(Man &m, std::vector<Stone> &s,Bonus &b, int jump_ongoing=0, double jumped=0.0, int anim_stone=0, double stone=0.0, int num = -1,  int pom_anim=0,int score_num=0)
  		: m_m(m), m_s(s),m_b(b), m_jump_ongoing(jump_ongoing), m_jumped(jumped), m_anim_stone(anim_stone), m_stone_speed(stone),  m_num(num),
                 m_pom_anim(pom_anim),m_score_num(score_num){
                     
@@ -41,8 +41,8 @@ public:
     int getPomAnim() const;
     void setPomAnim(int j);
     
-    float getScore() const ;
-    void setScore(float score);
+    int getScore() const ;
+    void setScore(int score);
     
 
 private:
